scm: tidy up non GUI dependent components of SCM interfaces
These components are primarily there to allow Patch Management (PM) tools
being used on top of an SCM controlled source tree to access file data from
the SCM point of view and make better decisions based on that.  E.g. if an
SCM tracked file is being added to a PM patch and it has changes from the
SCM POV then the PM can ask whether to include those changes in the patch
or not. But if it's a CLI tool it shouldn't have to depend on GUI stuff.

Signed-off-by: Peter Williams <pwil3058@gmail.com>